# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LamarIn API",
        "description": "REST API Sistem Informasi Lowongan Pekerjaan — Proyek Akhir Kelompok",
        "contact": {
            "email": "kelompok@lamarIn.dev"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Server Utama"
        }
    ],
    "paths": {
...
        "/api/lamaran": {
            "get": {
                "tags": [
                    "Lamaran"
                ],
                "summary": "Riwayat semua lamaran milik pelamar yang login",
                "operationId": "0ee6281ad49f4c1e6626e7fcabbc2bd3",
                "responses": {
                    "200": {
                        "description": "Daftar lamaran berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden — bukan pelamar"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Lamaran"
                ],
                "summary": "Kirim lamaran pada lowongan (hanya Pelamar)",
                "operationId": "32e2a4bc249c3f682e4a59735e372765",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "lowongan_id",
                                    "cv"
                                ],
                                "properties": {
                                    "lowongan_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "cv": {
                                        "description": "File CV (PDF/DOC/DOCX, maks 2MB)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "surat_lamaran": {
                                        "description": "Surat lamaran (opsional, PDF/DOC/DOCX, maks 2MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lamaran berhasil dikirim"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden — bukan pelamar"
                    },
                    "422": {
                        "description": "Validasi gagal atau lowongan tidak aktif"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/lamaran/{id}": {
            "get": {
                "tags": [
                    "Pelamar"
                ],
                "summary": "Melihat detail satu lamaran (hanya Penyedia pemilik lowongan)",
                "operationId": "1ccf7650853df9f40552d97d15fbf139",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Lamaran",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail lamaran berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Lamaran tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Lamaran"
                ],
                "summary": "Batalkan lamaran (hanya jika status masih 'menunggu')",
                "operationId": "1eb13d2c674df5f7610e604f5531d27d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lamaran berhasil dibatalkan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Lamaran tidak ditemukan"
                    },
                    "422": {
                        "description": "Lamaran tidak dapat dibatalkan karena sudah diproses"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}